Turn 'Full report' into a new module
Code has been entirely reviewed and is now simpler. A first .Rmd.origin
file is created with desired sections in code chunks. This file is then
rendered as a .Rmd file in which code has already be run. This file is
then render to get a HTML file which is used as a preview which can then
be print as PDF (see #11). All new sections are integrated (see #5).